+++ src/Common/com/bioxx/tfc/Blocks/Terrain/BlockDirt.java
@@ -151,7 +151,7 @@ public class BlockDirt extends BlockTerra
 
 			if(!canFallOneBelow && (count > 2) && sides.size() >= 1)
 			{
-				switch((Integer)sides.get(random.nextInt(sides.size())))
+				switch (sides.get(random.nextInt(sides.size())))
 				{
 				case 0:
 				{
