+++ src/Common/com/bioxx/tfc/Blocks/Terrain/BlockSand.java
@@ -144,7 +144,7 @@ public class BlockSand extends BlockTerra
 
 			if (!canFallOneBelow && (count > 2) && sides.size() >= 1)
 			{
-				switch((Integer)sides.get(random.nextInt(sides.size())))
+				switch (sides.get(random.nextInt(sides.size())))
 				{
 				case 0:
 				{
