+++ src/Common/com/bioxx/tfc/Entities/Mobs/EntityCowTFC.java
@@ -107,7 +107,7 @@ public class EntityCowTFC extends EntityCow implements IAnimal
 		//	in time such that they reach adulthood now.
 		//
 		//this.setGrowingAge((int) TFC_Time.getTotalDays() - getNumberOfDaysToAdult());
-		this.setAge((int) TFC_Time.getTotalDays() - getNumberOfDaysToAdult());
+		this.setAge(TFC_Time.getTotalDays() - getNumberOfDaysToAdult());
 		//For Testing Only(makes spawned animals into babies)
 		//this.setGrowingAge((int) TFC_Time.getTotalDays());
 	}
@@ -151,7 +151,7 @@ public class EntityCowTFC extends EntityCow implements IAnimal
 		//	We hijack the growingAge to hold the day of birth rather
 		//	than number of ticks to next growth event.
 		//
-		this.setAge((int) TFC_Time.getTotalDays());
+		this.setAge(TFC_Time.getTotalDays());
 	}
 
 	@Override
@@ -213,7 +213,7 @@ public class EntityCowTFC extends EntityCow implements IAnimal
 				baby.rotationYawHead = baby.rotationYaw;
 				baby.renderYawOffset = baby.rotationYaw;
 				worldObj.spawnEntityInWorld(baby);
-				baby.setAge((int)TFC_Time.getTotalDays());
+				baby.setAge(TFC_Time.getTotalDays());
 				pregnant = false;
 			}
 		}
